add an iterator for ByteArrayVec
